+++ packages/sdk-release-kit/src/cli/index.js
@@ -26,6 +26,7 @@ const args = require('yargs')
 const chalk = require('chalk')
 const cwd = process.cwd()
 const path = require('path')
+const fs = require('fs')
 const {verifyChangelog, writeReleaseEntryToChangelog} = require('../changelog')
 const {packInstall, lsDryRun} = require('../dry-run')
 const {lint} = require('../lint')
@@ -33,10 +34,12 @@ const sendReleaseNotification = require('../send-report')
 const {createDotFolder} = require('../setup')
 const {verifyCommits, verifyInstalledVersions, verifyVersions} = require('../versions')
 const {gitAdd, gitCommit, gitPullWithRebase, gitPushWithTags, isStagedForCommit} = require('../git')
+const {yarnInstall, yarnUpgrade, verifyUnfixedDeps} = require('../yarn')
+
+const command = args._[0]
 
 ;(async () => {
   try {
-    const command = args._[0]
     switch (command) {
       case 'lint':
       case 'l':
@@ -54,23 +57,38 @@ const {gitAdd, gitCommit, gitPullWithRebase, gitPushWithTags, isStagedForCommit}
       case 'preversion':
       case 'pre-version':
       case 'release-pre-check':
+        console.log('[bongo preversion] git pull')
         await gitPullWithRebase()
+        console.log('[bongo preversion] yarn install')
+        await yarnInstall()
+        console.log('[bongo preversion] yarn deps')
+        await deps()
+        console.log('[bongo preversion] lint')
         await lint(cwd)
+        console.log('[bongo preversion] verify changelog')
         await verifyChangelog(cwd)
-        if (!process.env.BONGO_SKIP_VERIFY_VERSIONS) {
-          await verifyVersions({isFix: args.fix, pkgPath: cwd})
+        console.log('[bongo preversion] verify unfixed dependencies')
+        verifyUnfixedDeps(cwd)
+        if (!args.skipVerifyVersions) {
+          console.log('[bongo preversion] verify commits')
+          await verifyCommits({pkgPath: cwd}).catch(err => console.log(err.message))
         }
-        if (!process.env.BONGO_SKIP_VERIFY_COMMITS) {
-          await verifyCommits({pkgPath: cwd})
+        try {
+          console.log('[bongo preversion] verify versions')
+          verifyVersions({pkgPath: cwd})
+        } catch (err) {
+          console.log(chalk.yellow(err.message))
         }
-        if (!process.env.BONGO_SKIP_VERIFY_INSTALLED_VERSIONS) {
+        if (!args.skipVerifyInstalledVersions) {
+          console.log('[bongo preversion] verify installed versions')
           createDotFolder(cwd)
           await packInstall(cwd)
-          return await verifyInstalledVersions({
+          await verifyInstalledVersions({
             pkgPath: cwd,
             installedDirectory: path.join('.bongo', 'dry-run'),
           })
         }
+        console.log('[bongo preversion] done!')
         return
       case 'send-release-notification':
       case 'hello-world':
@@ -83,10 +101,7 @@ const {gitAdd, gitCommit, gitPullWithRebase, gitPushWithTags, isStagedForCommit}
         return await verifyChangelog(cwd)
       case 'verify-commits':
       case 'vco':
-        if (!process.env.BONGO_SKIP_VERIFY_COMMITS) {
-          return await verifyCommits({pkgPath: cwd})
-        }
-        return
+        return await verifyCommits({pkgPath: cwd})
       case 'verify-installed-versions':
       case 'viv':
         createDotFolder(cwd)
@@ -97,25 +112,45 @@ const {gitAdd, gitCommit, gitPullWithRebase, gitPushWithTags, isStagedForCommit}
         })
       case 'verify-versions':
       case 'vv':
-        const isFix = args.fix
-        await verifyVersions({isFix, pkgPath: cwd})
-        if (isFix && !args.skipCommit) {
-          await gitAdd('package.json')
-          await gitAdd('CHANGELOG.md')
-          if (await isStagedForCommit('package.json', 'CHANGELOG.md')) {
-            await gitCommit()
-          }
+        try {
+          verifyVersions({pkgPath: cwd})
+        } catch (err) {
+          console.log(chalk.yellow(err.message))
         }
-        break
+        return
       case 'version':
         writeReleaseEntryToChangelog(cwd)
         return await gitAdd('CHANGELOG.md')
+      case 'deps':
+        return deps({shouldCommit: !args.skipCommit})
       default:
         throw new Error('Invalid option provided')
     }
   } catch (error) {
-    console.log(chalk.red(error.message))
-    console.log(error)
+    if (args.verbose) {
+      console.log(error)
+    } else {
+      console.log(chalk.red(error.message))
+      console.log(`run "npx bongo ${command} --verbose" to see stack trace`)
+    }
     process.exit(1)
   }
 })()
+
+async function deps({shouldCommit} = {}) {
+  verifyUnfixedDeps(cwd)
+  await yarnUpgrade({
+    folder: cwd,
+    upgradeAll: args.upgradeAll,
+    skipDev: args.skipDev,
+  })
+  if (shouldCommit) {
+    await gitAdd('package.json')
+    await gitAdd('CHANGELOG.md')
+    await gitAdd('yarn.lock')
+    if (await isStagedForCommit('package.json', 'CHANGELOG.md', 'yarn.lock')) {
+      const pkgName = JSON.parse(fs.readFileSync(path.resolve(cwd, 'package.json'))).name
+      await gitCommit(`[auto commit] ${pkgName}: upgrade deps`)
+    }
+  }
+}