+++ packages/eyes-protractor/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+- updated to @applitools/eyes-sdk-core@12.2.5 (from 12.2.4)
+- updated to @applitools/visual-grid-client@15.0.5 (from 15.0.4)
+- updated to @applitools/eyes-sdk-core@12.2.6 (from 12.2.5)
+- updated to @applitools/visual-grid-client@15.0.6 (from 15.0.5)
+
 
 ## 1.8.1 - 2020/9/15