+++ packages/eyes-webdriverio-5/CHANGELOG.md
@@ -3,6 +3,16 @@
 ## Unreleased
 
 
+## 5.23.1 - 2020/9/23
+
+- republish due to missing coverage tests
+
+## 5.23.0 - 2020/9/23
+
+- removed yarn workspaces
+- updated to @applitools/eyes-sdk-core@12.2.6 (from 12.2.5)
+- updated to @applitools/visual-grid-client@15.0.6 (from 15.0.5)
+
 ## 5.22.2 - 2020/9/17
 
 - updated to @applitools/eyes-sdk-core@12.2.5 (from 12.2.4)